LangChain-LocalAI/requirements.txt
langchain
openai
python-dotenv
sqlalchemy